# openapi/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "paths": {
  "/merge": {
   "post": {
    "summary": "Merge Pdfs",
    "description": "Merge passed page objects into a single PDF file. Only pages with checked=True will be merged.\n\nArgs:\n    request (List[Page]): List of Page objects to merge.\n\nReturns:\n    FileResponse: Merged PDF file.",
    "operationId": "merge_pdfs",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "items": {
         "$ref": "#/components/schemas/Page"
        },
        "type": "array",
        "title": "Request"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/upload": {
   "post": {
    "summary": "Upload File",
    "description": "Upload a PDF file to get all pages from that file.\n\nArgs:\n    file (UploadFile): The PDF file to be uploaded.\n\nReturns:\n    List[Page]: A list of page objects. extracted from the uploaded PDF file.",
    "operationId": "upload_file",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_file_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Page"
         },
         "type": "array",
         "title": "Response Upload File Upload Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "Body_upload_file_upload_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_file_upload_post"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "Page": {
    "properties": {
     "file_id": {
      "type": "string",
      "title": "File Id"
     },
     "page_number": {
      "type": "integer",
      "title": "Page Number"
     },
     "thumbnail": {
      "type": "string",
      "title": "Thumbnail"
     },
     "checked": {
      "type": "boolean",
      "title": "Checked",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "file_id",
     "page_number",
     "thumbnail"
    ],
    "title": "Page"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}